Refactor all the files.
