+++ algorithms/decision_tree.py
@@ -1,5 +1,8 @@
-import numpy as np
 from collections import Counter
+
+import numpy as np
+from sklearn import datasets
+from sklearn.model_selection import train_test_split
 
 
 def entropy(y):
@@ -113,3 +116,26 @@ class DecisionTree:
         counter = Counter(y)
         most_common = counter.most_common(1)[0][0]
         return most_common
+
+
+if __name__ == "__main__":
+
+    def accuracy(y_true, y_pred):
+        accuracy = np.sum(y_true == y_pred) / len(y_true)
+        return accuracy
+
+    data = datasets.load_breast_cancer()
+    X = data.data
+    y = data.target
+
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, test_size=0.2, random_state=1234
+    )
+
+    clf = DecisionTree(max_depth=10)
+    clf.fit(X_train, y_train)
+
+    y_pred = clf.predict(X_test)
+    acc = accuracy(y_test, y_pred)
+
+    print("Accuracy:", acc)
