+++ algorithms/kmeans.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib.pyplot as plt
+from sklearn.datasets import make_blobs
 
 np.random.seed(42)
 
@@ -98,3 +99,18 @@ class KMeans:
             ax.scatter(*point, marker="x", color="black", linewidth=2)
 
         plt.show()
+
+
+# Testing
+if __name__ == "__main__":
+    X, y = make_blobs(
+        centers=3, n_samples=500, n_features=2, shuffle=True, random_state=40
+    )
+    print(X.shape)
+
+    clusters = len(np.unique(y))
+    print(clusters)
+    k = KMeans(K=clusters, max_iters=150, plot_steps=True)
+    y_pred = k.predict(X)
+
+    k.plot()
