--- algorithms/lda_tests.py
@@ -1,29 +1,0 @@
-from sklearn import datasets
-import matplotlib.pyplot as plt
-import numpy as np
-
-from lda import LDA
-
-data = datasets.load_iris()
-X = data.data
-y = data.target
-
-# Project the data onto the 2 primary linear discriminants
-lda = LDA(2)
-lda.fit(X, y)
-X_projected = lda.transform(X)
-
-print("Shape of X:", X.shape)
-print("Shape of transformed X:", X_projected.shape)
-
-x1 = X_projected[:, 0]
-x2 = X_projected[:, 1]
-
-plt.scatter(
-    x1, x2, c=y, edgecolor="none", alpha=0.8, cmap=plt.cm.get_cmap("viridis", 3)
-)
-
-plt.xlabel("Linear Discriminant 1")
-plt.ylabel("Linear Discriminant 2")
-plt.colorbar()
-plt.show()
