+++ algorithms/logistic_regression.py
@@ -1,4 +1,7 @@
+import matplotlib.pyplot as plt
 import numpy as np
+from sklearn.model_selection import train_test_split
+from sklearn import datasets
 
 
 class LogisticRegression:
@@ -37,3 +40,24 @@ class LogisticRegression:
 
     def _sigmoid(self, x):
         return 1 / (1 + np.exp(-x))
+
+
+# Testing
+if __name__ == "__main__":
+
+    def accuracy(y_true, y_pred):
+        accuracy = np.sum(y_true == y_pred) / len(y_true)
+        return accuracy
+
+    bc = datasets.load_breast_cancer()
+    X, y = bc.data, bc.target
+
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, test_size=0.2, random_state=1234
+    )
+
+    regressor = LogisticRegression(learning_rate=0.0001, n_iters=1000)
+    regressor.fit(X_train, y_train)
+    predictions = regressor.predict(X_test)
+
+    print("LR classification accuracy:", accuracy(y_test, predictions))
