+++ algorithms/naivebayes.py
@@ -1,4 +1,7 @@
+import matplotlib.pyplot as plt
 import numpy as np
+from sklearn.model_selection import train_test_split
+from sklearn import datasets
 
 
 class NaiveBayes:
@@ -41,3 +44,24 @@ class NaiveBayes:
         numerator = np.exp(-((x - mean) ** 2) / (2 * var))
         denominator = np.sqrt(2 * np.pi * var)
         return numerator / denominator
+
+
+# Testing
+if __name__ == "__main__":
+
+    def accuracy(y_true, y_pred):
+        accuracy = np.sum(y_true == y_pred) / len(y_true)
+        return accuracy
+
+    X, y = datasets.make_classification(
+        n_samples=1000, n_features=10, n_classes=2, random_state=123
+    )
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, test_size=0.2, random_state=123
+    )
+
+    nb = NaiveBayes()
+    nb.fit(X_train, y_train)
+    predictions = nb.predict(X_test)
+
+    print("Naive Bayes classification accuracy", accuracy(y_test, predictions))
