+++ algorithms/random_forest.py
@@ -1,6 +1,10 @@
-import numpy as np
 from collections import Counter
-from decision_tree import DecisionTree
+
+import numpy as np
+from sklearn import datasets
+from sklearn.model_selection import train_test_split
+
+from .decision_tree import DecisionTree
 
 
 def bootstrap_sample(X, y):
@@ -40,3 +44,27 @@ class RandomForest:
         tree_preds = np.swapaxes(tree_preds, 0, 1)
         y_pred = [most_common_label(tree_pred) for tree_pred in tree_preds]
         return np.array(y_pred)
+
+
+# Testing
+if __name__ == "__main__":
+
+    def accuracy(y_true, y_pred):
+        accuracy = np.sum(y_true == y_pred) / len(y_true)
+        return accuracy
+
+    data = datasets.load_breast_cancer()
+    X = data.data
+    y = data.target
+
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, test_size=0.2, random_state=1234
+    )
+
+    clf = RandomForest(n_trees=3, max_depth=10)
+
+    clf.fit(X_train, y_train)
+    y_pred = clf.predict(X_test)
+    acc = accuracy(y_test, y_pred)
+
+    print("Accuracy:", acc)
